Revert "Swap section"
