+++ src/app/(home)/components/ComingSoon.tsx
@@ -1,56 +1,23 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 "use client";
-import React, { useState, useEffect, useRef, FormEventHandler } from "react";
+import React from "react";
 import { useTranslation } from "react-i18next";
-import { toast } from "react-hot-toast";
 
 import { renderText } from "@/utils/misc";
 import Image from "next/image";
 
 const ComingSoonSection = () => {
   const { t } = useTranslation("home");
-const [email, setEmail] = useState("");
-    const [loading, setLoading] = useState(false);
-  
-    const handleSubmit: FormEventHandler<HTMLFormElement> = async (e) => {
-      e.preventDefault();
-      setLoading(true);
-  
-      try {
-        const response = await fetch("/api/newsletter", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ email }),
-        });
-  
-        console.log(response);
-  
-        if (response.status === 200) {
-          toast.success("Thank you for subscribing!", {
-            style: { color: "#333" },
-          });
-          setEmail("");
-        } else {
-          toast.error("Something went wrong!", {
-            style: { color: "#333" },
-          });
-        }
-      } catch (error: any) {
-        console.error("Subscription error:", error);
-        toast.error(error?.response?.data?.message || "Unexpected error occurred.", {
-          style: { color: "#333" },
-        });
-      } finally {
-        setLoading(false);
-      }
-    };
+
 
 
   return (
-    <div>
-      <div className="mx-auto">
+
+    <div
+    // className={`w-full ${isAboutUs ? "bg-cta-gradient" : "bg-white"} py-20`}
+    >
+      <div className="mx-auto mt-20">
+
         <div className="relative bg-primary-gradient p-2 md:p-8 md:p-12">
           <div className="relative mx-auto max-w-7xl">
             <div className="img1 md:absolute top-10 left-0 mb-2 md:mb-0">
@@ -89,26 +56,7 @@ const [email, setEmail] = useState("");
                     alt="phone-apps-screenshots"
                   />
                 </div>
-  <form
-                  onSubmit={handleSubmit}
-                  className="flex flex-col md:flex-row  w-full md:min-w-[500px] justify-center md:w-2/4 mt-10 items-center gap-4 sm:mb-0 md:bg-[#9794D8] px-2 pl-4 py-2 rounded-full"
-                >
-                  <input
-                    type="email"
-                    name="email"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    placeholder={t("coming-soon.waitlist")}
-                    className="bg-[#9794D8] px-4 py-2 md:p-2 rouneded-lg w-full md:w-auto border-white md:border-0 md:bg-transparent flex flex-1 text-white placeholder-white outline-none border-0 rounded-full focus:ring-0 focus:outline-none"
-                    required
-                  />
-                  <button
-                    type="submit"
-                    className="bg-white w-auto hover:bg-gray-200 text-blue-custom-700 px-6 py-2 rounded-full md:font-semibold transition-colors"
-                  >
-                    {loading ? renderText(t("Loading...")) : renderText(t("join"))}
-                  </button>
-                </form>
+
               <br />
               </div>
             </div>
