+++ src/app/(home)/components/Values.tsx
@@ -1,9 +1,9 @@
 "use client";
 import { renderText } from "@/utils/misc";
 import Link from "next/link";
-import React, { useState, useEffect, useRef} from "react";
+import React, { useState, useEffect, useRef, FormEventHandler } from "react";
 import { useTranslation } from "react-i18next";
-
+import { toast } from "react-hot-toast";
 const useCounter = (end: number, duration = 10000, shouldStart: boolean) => {
   const [count, setCount] = useState(0);
   const countRef = useRef(0);
@@ -74,7 +74,43 @@ const useInView = () => {
 };
 
 const ValueStats = () => {
-   
+   const [email, setEmail] = useState("");
+    const [loading, setLoading] = useState(false);
+  
+    const handleSubmit: FormEventHandler<HTMLFormElement> = async (e) => {
+      e.preventDefault();
+      setLoading(true);
+  
+      try {
+        const response = await fetch("/api/newsletter", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ email }),
+        });
+  
+        console.log(response);
+  
+        if (response.status === 200) {
+          toast.success("Thank you for subscribing!", {
+            style: { color: "#333" },
+          });
+          setEmail("");
+        } else {
+          toast.error("Something went wrong!", {
+            style: { color: "#333" },
+          });
+        }
+      } catch (error: any) {
+        console.error("Subscription error:", error);
+        toast.error(error?.response?.data?.message || "Unexpected error occurred.", {
+          style: { color: "#333" },
+        });
+      } finally {
+        setLoading(false);
+      }
+    };
   const { t } = useTranslation("home");
 
   const [ref, isInView] = useInView();
@@ -122,7 +158,26 @@ const ValueStats = () => {
           </Link>   
         </div>
           <div className="flex justify-center items-center">
-          
+            <form
+                  onSubmit={handleSubmit}
+                  className="flex flex-col md:flex-row  w-full md:min-w-[500px] justify-center md:w-2/4 mt-10 items-center gap-4 sm:mb-0 md:bg-[#9794D8] px-2 pl-4 py-2 rounded-full"
+                >
+                  <input
+                    type="email"
+                    name="email"
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}
+                    placeholder={t("coming-soon.waitlist")}
+                    className="bg-[#9794D8] px-4 py-2 md:p-2 rouneded-lg w-full md:w-auto border-white md:border-0 md:bg-transparent flex flex-1 text-white placeholder-white outline-none border-0 rounded-full focus:ring-0 focus:outline-none"
+                    required
+                  />
+                  <button
+                    type="submit"
+                    className="bg-white w-auto hover:bg-gray-200 text-blue-custom-700 px-6 py-2 rounded-full md:font-semibold transition-colors"
+                  >
+                    {loading ? renderText(t("Loading...")) : renderText(t("join"))}
+                  </button>
+                </form>
           </div>
       </div>
     </section>
